refactor: update imports to use absolute paths for localPartyId and related components
- Replaced relative imports of localPartyId and mpcLocalPartyId with absolute paths across multiple files for improved clarity and consistency.
- Introduced new utility functions in localPartyId to enhance local party ID generation and parsing.
- Updated components to utilize the new structure, ensuring seamless integration with the updated import paths.
