+++ clients/desktop/src/mpc/peers/option/PeerOption.tsx
@@ -1,3 +1,7 @@
+import {
+  formatMpcDeviceName,
+  parseLocalPartyId,
+} from '@core/mpc/signers/localPartyId'
 import { centerContent } from '@lib/ui/css/centerContent'
 import { round } from '@lib/ui/css/round'
 import { sameDimensions } from '@lib/ui/css/sameDimensions'
@@ -8,7 +12,6 @@ import { Text } from '@lib/ui/text'
 import { getColor } from '@lib/ui/theme/getters'
 import styled from 'styled-components'
 
-import { formatMpcDeviceName, parseLocalPartyId } from '../../localPartyId'
 import { useMpcPeersSelectionRecord } from '../state/mpcSelectedPeers'
 import { PeerOptionContainer } from './PeerOptionContainer'
