+++ clients/desktop/src/vault/keygen/shared/mutations/useDklsKeygen.ts
@@ -19,7 +19,7 @@ import {
   GetLocalUIEcdsa,
   GetLocalUIEdDSA,
 } from '../../../../../wailsjs/go/tss/TssService'
-import { useMpcLocalPartyId } from '../../../../mpc/localPartyId/state/mpcLocalPartyId'
+import { useMpcLocalPartyId } from '@core/ui/mpc/state/mpcLocalPartyId'
 import { useMpcPeers } from '../../../../mpc/peers/state/mpcPeers'
 import { useMpcServerUrl } from '../../../../mpc/serverType/state/mpcServerUrl'
 import { useMpcSessionId } from '../../../../mpc/session/state/mpcSession'
