+++ clients/desktop/src/vault/keygen/shared/peerDiscovery/KeygenPeerDiscoveryStep.tsx
@@ -1,6 +1,8 @@
 import { KeygenType } from '@core/mpc/keygen/KeygenType'
 import { recommendedPeers, requiredPeers } from '@core/mpc/peers/config'
+import { parseLocalPartyId } from '@core/mpc/signers/localPartyId'
 import { useKeygenVault } from '@core/ui/mpc/keygen/state/keygenVault'
+import { useMpcLocalPartyId } from '@core/ui/mpc/state/mpcLocalPartyId'
 import { Match } from '@lib/ui/base/Match'
 import { InfoIcon } from '@lib/ui/icons/InfoIcon'
 import { OnBackProp, OnForwardProp } from '@lib/ui/props'
@@ -14,8 +16,6 @@ import { useTranslation } from 'react-i18next'
 
 import { getFormProps } from '../../../../lib/ui/form/utils/getFormProps'
 import { QueryBasedQrCode } from '../../../../lib/ui/qr/QueryBasedQrCode'
-import { parseLocalPartyId } from '../../../../mpc/localPartyId'
-import { useMpcLocalPartyId } from '../../../../mpc/localPartyId/state/mpcLocalPartyId'
 import { InitiatingDevice } from '../../../../mpc/peers/InitiatingDevice'
 import { PeerOption } from '../../../../mpc/peers/option/PeerOption'
 import { PeerDiscoveryFormFooter } from '../../../../mpc/peers/PeerDiscoveryFormFooter'
