+++ clients/desktop/src/vault/keygen/shared/peerDiscovery/queries/usePeerOptionsQuery.ts
@@ -4,7 +4,7 @@ import { withoutDuplicates } from '@lib/utils/array/withoutDuplicates'
 import { queryUrl } from '@lib/utils/query/queryUrl'
 import { useQuery } from '@tanstack/react-query'
 
-import { useMpcLocalPartyId } from '../../../../../mpc/localPartyId/state/mpcLocalPartyId'
+import { useMpcLocalPartyId } from '@core/ui/mpc/state/mpcLocalPartyId'
 import { useMpcServerUrl } from '../../../../../mpc/serverType/state/mpcServerUrl'
 import { useMpcSessionId } from '../../../../../mpc/session/state/mpcSession'
