+++ clients/desktop/src/vault/reshare/fast/FastReshareServerStep.tsx
@@ -1,10 +1,10 @@
+import { generateLocalPartyId } from '@core/mpc/signers/localPartyId'
 import { OnForwardProp } from '@lib/ui/props'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { useMutation } from '@tanstack/react-query'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { generateLocalPartyId } from '../../../mpc/localPartyId'
 import { useMpcSessionId } from '../../../mpc/session/state/mpcSession'
 import { FullPageFlowErrorState } from '../../../ui/flow/FullPageFlowErrorState'
 import { PageHeader } from '../../../ui/page/PageHeader'
