+++ clients/desktop/src/vault/reshare/shared/ReshareVerifyStep.tsx
@@ -12,7 +12,7 @@ import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
 
 import { InfoBlock } from '../../../lib/ui/status/InfoBlock'
-import { useMpcLocalPartyId } from '../../../mpc/localPartyId/state/mpcLocalPartyId'
+import { useMpcLocalPartyId } from '@core/ui/mpc/state/mpcLocalPartyId'
 import { useMpcSigners } from '../../../mpc/signers/state/mpcSigners'
 import { FlowPageHeader } from '../../../ui/flow/FlowPageHeader'
 import { PageContent } from '../../../ui/page/PageContent'
